Refactor interaction api
