+++ src/commands/course/subcommands/delete.ts
@@ -22,13 +22,15 @@ const del = new SubCommand({
     const course = await ctx.courses().getByRoleId(guildId, roleId);
 
     if (!course) {
-      return ctx.respondWithError('That course does not exist');
+      return ctx.interactionApi.respondWithError('That course does not exist');
     }
 
     await ctx.api.deleteGuildRole(guildId, course.roleId);
     await ctx.courses().deleteByRoleId(guildId, course.roleId);
 
-    return ctx.respondWithMessage(`Deleted course ${boldCourse(course)}`);
+    return ctx.interactionApi.respondWithMessage(
+      `Deleted course ${boldCourse(course)}`,
+    );
   },
 });
