+++ src/commands/course/subcommands/leave.ts
@@ -21,21 +21,23 @@ const leave = new SubCommand({
 
     const course = await ctx.courses().getByRoleId(guildId, roleId);
     if (!course) {
-      return ctx.respondWithError('That course does not exist');
+      return ctx.interactionApi.respondWithError('That course does not exist');
     }
 
     const userId = ctx.mustGetUserId();
 
     const members = (await ctx.courses().getMembers(guildId, roleId)) ?? [];
     if (!members.includes(userId)) {
-      return ctx.respondWithError(`You aren't in that course`);
+      return ctx.interactionApi.respondWithError(`You aren't in that course`);
     }
 
     // `GUILD_MEMBER_UPDATE` trigger will activate and automatically remove the
     // user from the database. Thus, there's no need to do it here.
     await ctx.api.removeRoleFromMember(guildId, userId, roleId);
 
-    return ctx.respondWithMessage(`You left the course ${boldCourse(course)}`);
+    return ctx.interactionApi.respondWithMessage(
+      `You left the course ${boldCourse(course)}`,
+    );
   },
 });
