+++ src/commands/course/subcommands/listJoined.ts
@@ -83,7 +83,7 @@ const listJoined = new SubCommand({
       value: sub.list, // The course list.
     }));
 
-    await ctx.respondSilentlyWithEmbed({
+    await ctx.interactionApi.respondSilentlyWithEmbed({
       type: Discord.EmbedType.RICH,
       title: `${username}’s Course List`,
       fields,
