+++ src/commands/ping.ts
@@ -7,13 +7,13 @@ const ping = new Command({
   handler: async ctx => {
     const startTime = ctx.interactionDate();
 
-    await ctx.respondWithMessage('pong');
+    await ctx.interactionApi.respondWithMessage('pong');
 
-    const msg = await ctx.getResponse();
+    const msg = await ctx.interactionApi.getResponse();
     const endTime = new Date(msg.timestamp);
     const latency = endTime.getTime() - startTime.getTime();
 
-    await ctx.editResponse({
+    await ctx.interactionApi.editResponse({
       content: `pong (latency: ${latency}ms)`,
     });
   },
